+++ Kernel/src/storage/partitiondevice.cpp
@@ -6,7 +6,7 @@ PartitionDevice::PartitionDevice(uint64_t startLBA, uint64_t endLBA, DiskDevice*
     this->parentDisk = disk;
 }
 
-int PartitionDevice::Read(uint64_t lba, uint16_t count, void* buffer){
+int PartitionDevice::Read(uint64_t lba, uint32_t count, void* buffer){
     if(count * parentDisk->blocksize > (endLBA - startLBA) * parentDisk->blocksize) return 2;
 
     return parentDisk->Read(lba + startLBA, count, buffer);
